# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/v1/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of User",
                "description": "Returns list of User",
                "operationId": "getUserList",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/users?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/users?page=1"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "url": null,
                                                    "label": "&laquo; Previous",
                                                    "active": false
                                                },
                                                {
                                                    "url": "http://localhost/api/v1/users?page=1",
                                                    "label": "1",
                                                    "active": true
                                                },
                                                {
                                                    "url": null,
                                                    "label": "Next &raquo;",
                                                    "active": false
                                                }
                                            ]
                                        },
                                        "next_page_url": {
                                            "type": "integer",
                                            "example": "null"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/users"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": "15"
                                        },
                                        "prev_page_url": {
                                            "type": "integer",
                                            "example": "null"
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "description": "Add User",
                "operationId": "addUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "$ref": "#/components/schemas/User/properties/name"
                                    },
                                    "email": {
                                        "$ref": "#/components/schemas/User/properties/email"
                                    },
                                    "password": {
                                        "$ref": "#/components/schemas/User/properties/password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/v1/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User By Id",
                "description": "Returns updated User data",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update existing User",
                "description": "Returns updated User data",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/User/properties/id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "$ref": "#/components/schemas/User/properties/name"
                                    },
                                    "email": {
                                        "$ref": "#/components/schemas/User/properties/email"
                                    },
                                    "password": {
                                        "$ref": "#/components/schemas/User/properties/password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete User",
                "description": "Returns updated User data",
                "operationId": "deleteUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User deleted."
                                        },
                                        "deleted": {
                                            "type": "bool",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "minimum": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string",
                        "format": "email"
                    },
                    "emailVerifiedAt": {
                        "title": "email_verified_at",
                        "description": "email_verified_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string"
                    },
                    "rememberToken": {
                        "title": "remember_token",
                        "description": "remember_token",
                        "type": "string"
                    },
                    "createdAt": {
                        "title": "created_at",
                        "description": "created_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "title": "updated_at",
                        "description": "updated_at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "401": {
                "description": "Error: Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "403": {
                "description": "Error: Forbidden",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "This action is unauthorized."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "404": {
                "description": "Resource Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Record not found."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "422": {
                "description": "Error: Unprocessable Entity",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object"
                        },
                        "example": {
                            "message": "The given data was invalid.",
                            "errors": {
                                "name": [
                                    "The name field is required."
                                ]
                            }
                        }
                    }
                }
            }
        },
        "parameters": {
            "page": {
                "name": "page",
                "in": "query",
                "description": "data page",
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}